+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## `0.9.0` Upcoming
+## `0.9.0`
 
 - `#80`: As dev, I want to have build/test/deploy status badge on `README.md`
 - `#35`: As scp user, I want to choose skip download if local file has bigger size. (`option.skipIfBigger` default to `false`)
@@ -8,6 +8,7 @@
 - `#85`: bug: As scp user, I want to remove mutual exclusive check on file size filters.
 - `#87`: bug: As scp user, I want to remove mutual exclusive check on file modified time filters.
 - `#84`: As scp user, I want to choose skip download if file sizes are same. (`option.skipIfSameSize` default to `false`)
+- `#86`: As scp user, I want to choose skip download if file modified time are same. (`option.skipIfSameAge` default to `false`)
 
 ## `0.8.0`